fix: 画面のデザイン修正とリファクタリング
